Reduce requirement for multiple presentations
Reduce requirement for multiple presentations from the same group in one semester.  It was previously decided that we'd be doing one presentation per group per semester because one of the presentations was always redundant as well as not having enough bandwidth to have everyone present twice.

We should change this requirement to reflect our previous decision.